pyproject.toml
[tool.poetry]
name = "dj_sinp_organisms"
version = "1.1.0"
description = "Django app to manage french SINP organisms standard"
authors = ["dbChiro project <project@dbchiro.org>"]
license = "AGPLv3"
keywords = ["SINP", "Nomenclatures", "Django", "France", "dbChiroWeb"]
readme = "README.rst"
homepage = "https://github.com/dbchiro/DjangoSinpNomenclature"
repository = "https://github.com/dbchiro/DjangoSinpNomenclature"
include = ["LICENSE"]
packages = [{ include = "sinp_organisms" }]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
Django = ">=3.2"
djangorestframework = "^3"
dj-sinp-nomenclatures = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest = "^7.4.3"
isort = "^5.13.2"
flake8 = ">=6.1,<8.0"
flake8-isort = "^6.1.1"
pre-commit = "^3.5"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
dj-database-url = "^2.1.0"
#dj-sinp-nomenclatures = { path = "../dj_sinp_nomenclatures", develop = true }
bpython = "^0.24"
django-fixture-magic = "^0.1.5"


[tool.poetry.group.docs.dependencies]
sphinx = "^6"
sphinx-rtd-theme = "^2.0.0"
rstcheck = "^6.2.0"
myst-parser = "^2.0.0"

[tool.isort]
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
include = '\.pyi?$'
line-length = 79
target-version = ['py38']


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
